Add memory limit, correct tmp dir remove rule

+++ process/l8-combine-rgb.sh
@@ -1,24 +1,24 @@
 # /bin/bash
 
-if [ "$#" -ne 1 ]
+if [ "$#" -ne 2 ]
 then
   echo "    Usage: $0 [File name]"
-  echo "  Example: $0 /tmp/LC81180442014356LGN00/final/final-rgb.TIF"
+  echo "  Example: $0 /path-to/final-rgb-pan.TIF /tmp/LC81180442014356LGN00/final/final-rgb.TIF"
   exit 1
 fi
 
-FINAL=$1
-FILENAME=`basename $1`
+FINAL=$2
+FILENAME=`basename $2`
 NAME=${FILENAME%.*}
-DIR=`dirname $1`
+DIR=`dirname $2`
 TMP="${DIR}/tmp"
 mkdir -p $TMP
 
-cp $1 $TMP/
+cp $1 $TMP/rgb-pan.tif
 
 # processing image
-convert -monitor -channel B -gamma 0.98 -channel R -gamma 1.03 -sigmoidal-contrast 30x15% $TMP/$FILENAME $TMP/rgb.tif
-cp -f $TMP/rgb.tif $FINAL
+convert -monitor -channel B -gamma 0.98 -channel R -gamma 1.03 -sigmoidal-contrast 30x15% $TMP/rgb-pan.tif $TMP/rgb-pan-light.tif
+cp -f $TMP/rgb-pan-light.tif $FINAL
 
 # append geotiff into image
 if [ -f $DIR/${NAME}.tfw ]; then
@@ -26,4 +26,6 @@ if [ -f $DIR/${NAME}.tfw ]; then
 fi
 
 # clean up
-rm -Rf $TMP
+if [ -f $FINAL ]; then
+  rm -Rf $TMP
+fi
